* Replace shell/make scripts use for svn_version.h generation by use src\tools\gensvnrevision (used at Windows). Note: --enable-svn-extract configure option removed, src\tools\gensvnrevision proper work in all cases without this option support. * Fixed: some GCC warnings.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@6674 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/shared/Database/DBCStores.cpp
@@ -202,7 +202,7 @@ void LoadDBCStores(std::string dataPath)
         if(AreaTableEntry const* area = sAreaStore.LookupEntry(i))
         {
             // fill AreaId->DBC records
-            sAreaFlagByAreaID.insert(AreaFlagByAreaID::value_type(area->ID,area->exploreFlag));
+            sAreaFlagByAreaID.insert(AreaFlagByAreaID::value_type(uint16(area->ID),area->exploreFlag));
 
             // fill MapId->DBC records ( skip sub zones and continents )
             if(area->zone==0 && area->mapid != 0 && area->mapid != 1 && area->mapid != 530 )
